http: Add error handling for the repo-locked exception. Part of #4237
Previously the repo locekd exception was raised on return
of the call to the hooks daemon and propagates all the
way up to the WSGI server (e.g. waitress). In case of
locked repos we want to return a custom response which
contains an explanation for the user that the repo is
locked and who has locked it.