nix: Use --set instead of --prefix for setting the PATH / PYTHONPATH
Using --set will not append the system PATH / PYTHONPATH to the
environment variable during wraping of executables. This will "lock-out"
programs and python modules from the system and only allow the ones
which are set during wrapping.

This means e.g. of git we make only our custom git available and
if thats missing we get an error instead of silently using the system
git (if installed)